Search Page Styling Update
Restyled the search page so that it showed up as a carousel of books with a focus on the images of the books rather than a long list, keeps the style uniform with what we have on the discover page. In the future, I'd like to layer the carousels by genre from top to bottom.